#2245 Accordion animatie verspringt
